refactor(cli): extract workflow stress handler
